Fix UI tests
